started fixing object layers

+++ Source/MonoGame.Extended.Content.Pipeline/Tiled/TmxObject.cs
@@ -14,20 +14,20 @@ namespace MonoGame.Extended.Content.Pipeline.Tiled
         [XmlAttribute(DataType = "string", AttributeName = "type")]
         public string Type { get; set; }
 
-        [XmlAttribute(DataType = "int", AttributeName = "x")]
-        public int X { get; set; }
+        [XmlAttribute(DataType = "float", AttributeName = "x")]
+        public float X { get; set; }
 
-        [XmlAttribute(DataType = "int", AttributeName = "y")]
-        public int Y { get; set; }
+        [XmlAttribute(DataType = "float", AttributeName = "y")]
+        public float Y { get; set; }
 
-        [XmlAttribute(DataType = "int", AttributeName = "width")]
-        public int Width { get; set; }
+        [XmlAttribute(DataType = "float", AttributeName = "width")]
+        public float Width { get; set; }
 
-        [XmlAttribute(DataType = "int", AttributeName = "height")]
-        public int Height { get; set; }
+        [XmlAttribute(DataType = "float", AttributeName = "height")]
+        public float Height { get; set; }
 
-        [XmlAttribute(DataType = "int", AttributeName = "rotation")]
-        public int Rotation { get; set; }
+        [XmlAttribute(DataType = "float", AttributeName = "rotation")]
+        public float Rotation { get; set; }
         
         [XmlAttribute(DataType = "boolean", AttributeName = "visible")]
         public bool Visible { get; set; }
